Updating plugin descriptor parsing code to not require child elements of the config when the format is raw.
The plugin parsing code previously threw an exception if a resource type in the plugin descriptor declared
<resource-configuration/> without any child elements. Those children correspond to structured configuration.
For a resource type that supports only raw, there is no need to specify any child elements. The parsing code
has been updated to support this.

Updating raw-config-test plugin with a second resource type that supports only raw configuration which is
how I discovered the parsing behavior.