[US 4] Fixing bugs in the merge() method that resulted in returning the configuration unmodified.
In addition to returning the updated configuration, the merge() method has been updated to "refresh" the side
being merged into. If for example, we are merging from structured to raw, ConfigurationManager first loads the
raw configs from config facet, and replaces Configuration.rawConfigurations with those loaded raw configs.
Then the structured is merged into the raws. The same is done going from raw to structured.