Initial commit for raw config ui changes
At this point thing raw config ui is broken as I am trying to learn how to rebuild the UI
with faclets and richfaces. I have the config files rendering in a rich panel menu, and
selecting a file in the menu will cause its contents to be displayed in the text area.
Selecting the file will also cause the path name to change and appear above the text area
on the upper left. The download functionality has been pushed into a backing bean method.

Updating raw-config-test plugin so that RawServer supports multiple raw files now.