[BZ 601153] Modifying configu update logic to avoid TransientObjectException
During the resource config update code, the parent reference of a Property was getting
set to a new (i.e., transient) Configuration object. Because of the delete_orphan cascade
style on the Property map, this was resulting in a TransientObjectException.

Also updating exception handling logic in ApacheServerComponent.updateResourceConfiguration
to capture the augeas error message if present in the config update report.