Refactoring the way plugin container can be started up in the tests and converted the resource upgrade tests to use this new approach.
When the test inherits from PluginContainerTest or uses it as an TestNG
listener, the plugin container can be configured declaratively using
an @PluginContainerSetup annotation. There is a couple of static methods
on the PluginContainerTest class to work with such preconfigured PC
(like starting it).

The JMockTest generic base class was augmented so that it can be used as both
a base class to tests and a TestNG listener.