[BZ 788638] Removing extra call to fetch managed object
We were making an extra call to getManagedCompoment in
AbstractEjbComponent during metric collection. I discovered this while
doing some profiling. With this change execution time for
AbstractEjbBeanComponent.getValues is reduced by about 50% Also adding
some logging in ManagedComponentComponent.getAvailability to indicate
when an availability refresh is needed. This could be useful for tuning
the interval.