[BZ 846400] Add validation for components that rely connectors and discovery group name. A compoment should have either the connector or discovery group configured, but not both.
This validation is handled by the AS7 server itself for new resource creation but the validation is missing for configuration updates. Without this validation code in place, the AS7 server accepts the commands, persists the configuration, but fails to reload or restart rendering the server unusable until the configuration file is manually edited. With this fix, the AS7 plugin will not send the configuration updates to the AS7 server unless they meet the descriptor requirements.