Adding initial support for ColumnFamily resource type
Cassandra's JMX API does not expose an attribute or operation for retrieving column
families; so, with this commit I discover column families using Hector. The JMX API does
expose an attribute to obtain a list of keyspaces, but I am unaware of any attributes or
operations to get a list of column families.

This commit just handles discovery and setting the objectName plugin config property of the
column family resource component.