Adding repair operation for StorageService type
The Cassandra docs recommend running repair as a routine maintenance operation. See
http://wiki.apache.org/cassandra/Operations#Frequency_of_nodetool_repair for details.
Interestingly the JMX operations for repair are exposed on the storage service mbean. The
frequencey of how often it is run should be determined in part by the value of
gc_grace_seconds which is specified at the column family level.

From a user standpoint, it will be somewhat difficult to determine that minimum
frequency as it will require querying the gc_grace_seconds config property for each column
family; however, that property will likely remain the same across column families. We could
expose the repair operation at the column family level, using a safe default interval
<= gc_grace_seconds. This could result in multiple repairs running simultaneously. I am not
sure whether or not this would/could be problematic.